Skip TLS tests when TLS memcached is not available
You can get a TLS enabled memcached by compiling it with --enable-tls.
